requirements.txt
pymysql
lxml
Flask
SQLAlchemy
SQLAlchemy-Utils
Flask-SQLAlchemy
Flask-Testing
Flask-Principal
Flask-Session
Flask-Script
Flask-Migrate
marshmallow
webargs
redis
phonenumbers
simplejson
dkimpy
dnspython
PySocks
requests
pycrypto
colorlog
nose
gunicorn
arrow
requests-mock
Faker
coverage
elasticsearch-dsl
user_agents
jira
redlock-py == 1.0.8
